Media Query and a lot of other stuff
I did a lot here without committing, but I'll try to explain. I'm still experimenting with css so it might be a bit jumbled. I added new images and added a sticky element to the images to get them to stick to the top of the page while you scroll. I added media queries to reflect a similar idea at <=600px.